useAttachmentDownload hook
